+++ packages/config/src/index.ts
@@ -2,3 +2,4 @@ export * from "./env";
 export * from "./constants";
 export * from "./logger";
 export * from "./indonesian_payroll_engine";
+export * from "./pay-structure-payroll";